[refactor] move some logic-related code
